+++ projects/can_transceiver/src/can_transceiver_ros_intf.cpp
@@ -1,3 +1,5 @@
+#include <memory>
+
 #include "can_frame_parser.h"
 #include "can_transceiver.h"
 #include "rclcpp/publisher.hpp"
@@ -47,6 +49,50 @@ private:
      */
     void sub_cb(std_msgs::msg::String /* placeholder */) { can_trns_->onNewCmd(Placeholder0); }
 };
+
+//===========================================================================================
+// Simulation Interface Components
+//===========================================================================================
+// [Simulator]<->[CAN Simulation Interface]<->[Virtual CAN bus]<->[CAN Transceiver]
+
+// Refer to https://ubcsailbot.atlassian.net/wiki/spaces/prjt22/pages/1768849494/Simulation+Interface#Interfaces
+
+/**
+ * Implementation of CAN Transceiver that interfaces with the simulator
+ * This replaces the CAN bus during testing to interface with simulator
+ *
+ * Receives info from ROS node with topic, specified by controls team
+ * Refer to https://ubcsailbot.atlassian.net/wiki/spaces/prjt22/pages/1785790595/Boat+Simulator+Design+Specification
+ * 
+ * Receives from ROS topic, /Simulator. Controls team sends to that
+ * This is done by subscribing to ROS node /Simulator
+ */
+
+class CanSimIntf : public CanTransceiver, public rclcpp::Node
+{
+    // There is no timer because subscriber will respond to whatever data is published to the topic /Simulator
+public:
+    CanSimIntf() : Node("Simulator")
+    {
+        int simulatorData = 0;  //Placeholder
+
+        subscription_ = this->create_subscription<std_msgs::msg::String>(
+          // Topic is called /Simulator
+          "Simulator", simulatorData, std::bind(&CanSimIntf::topic_callback, this, std::placeholders::_1));
+    }
+
+private:
+    // Receives string message data over topic and writes to RCLCPP_INFO macro
+    void topic_callback(const std_msgs::msg::String::SharedPtr msg) const
+    {
+        RCLCPP_INFO(this->get_logger(), "I heard: '%s'", msg->data.c_str());
+    }
+    rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
+};
+
+//===========================================================================================
+// Main
+//===========================================================================================
 
 int main(int argc, char * argv[])
 {
